requirements.txt
python-dotenv
scrapy
beautifulsoup4
pandas
pymongo
sentence_transformers
streamlit
openai==0.28
flask
flask-cors
scikit-learn
regex
